+++ ark_nlp/factory/task/base/__init__.py
@@ -1,3 +1,4 @@
 from ark_nlp.factory.task.base._task import Task
+from ark_nlp.factory.task.base._task_mixin import TaskMixin
 from ark_nlp.factory.task.base._token_classification import TokenClassificationTask
 from ark_nlp.factory.task.base._sequence_classification import SequenceClassificationTask
