+++ ark_nlp/model/ie/prompt_uie/task.py
@@ -87,11 +87,11 @@ class PromptUIETask(TokenClassificationTask):
 
         return None
 
-    def on_evaluate_step_end(self, inputs, logits, **kwargs):
+    def on_evaluate_step_end(self, inputs, outputs, **kwargs):
 
         with torch.no_grad():
             # compute loss
-            logits, loss = self._get_evaluate_loss(inputs, logits, **kwargs)
+            logits, loss = self._get_evaluate_loss(inputs, outputs, **kwargs)
 
         S = []
         start_logits = logits[0]
